+++ src/mytester.php
@@ -7,6 +7,11 @@ $vendorDirectory = findVendorDirectory();
 
 require $vendorDirectory . "/autoload.php";
 
+use MyTester\CodeCoverage\Collector;
+use MyTester\CodeCoverage\Helper as CodeCoverageHelper;
+use MyTester\CodeCoverage\PcovEngine;
+use MyTester\CodeCoverage\PercentFormatter;
+use MyTester\CodeCoverage\XDebugEngine;
 use MyTester\Tester;
 use Nette\CommandLine\Parser;
 
@@ -17,9 +22,24 @@ $cmd = new Parser("", [
     "--colors" => [
         Parser::Optional => true,
     ],
+    "--coverageFormat" => [
+        Parser::Argument => true,
+        Parser::Optional => true,
+        Parser::Enum => array_keys(CodeCoverageHelper::$availableFormatters),
+    ],
 ]);
 $options = $cmd->parse();
 
-$tester = new Tester($options["path"]);
+$codeCoverageCollector = new Collector();
+$codeCoverageCollector->registerEngine(new PcovEngine());
+$codeCoverageCollector->registerEngine(new XDebugEngine());
+$codeCoverageCollector->registerFormatter(new PercentFormatter());
+$coverageFormat = $options["--coverageFormat"];
+if ($coverageFormat !== null) {
+    $type = CodeCoverageHelper::$availableFormatters[$coverageFormat];
+    $codeCoverageCollector->registerFormatter(new $type());// @phpstan-ignore argument.type
+}
+
+$tester = new Tester(folder: $options["path"], codeCoverageCollector: $codeCoverageCollector);
 $tester->useColors = isset($options["--colors"]);
 $tester->execute();